refactor relativeTime function
