+++ src/routes/admin/connections/+page.svelte
@@ -4,6 +4,7 @@
   import * as Table from "$lib/components/ui/table";
   import type { PageData } from "./$types";
   import type { Connection, User } from "@prisma/client";
+  import { relativeTime } from "$lib/date";
 
   export let data: PageData;
 
@@ -42,29 +43,6 @@
 
   const { headerRows, pageRows, tableAttrs, tableBodyAttrs } =
     table.createViewModel(columns);
-
-  function relativeTime(fromDate: Date, toDate: Date = new Date()) {
-    const units = [
-      { max: 1000 * 60, name: "second", divisor: 1000 },
-      { max: 1000 * 60 * 60, name: "minute", divisor: 1000 * 60 },
-      { max: 1000 * 60 * 60 * 24, name: "hour", divisor: 1000 * 60 * 60 },
-      {
-        max: 1000 * 60 * 60 * 24 * 30,
-        name: "day",
-        divisor: 1000 * 60 * 60 * 24,
-      },
-    ];
-
-    const millis = fromDate.getTime() - toDate.getTime();
-
-    const unit = units.find((u) => Math.abs(millis) < u.max);
-    if (!unit) return "Unknown";
-
-    const difference = Math.round(millis / unit.divisor);
-
-    const rtf = new Intl.RelativeTimeFormat("en", { numeric: "auto" });
-    return rtf.format(difference, unit.name as any);
-  }
 </script>
 
 <div class="flex items-center justify-between">
